--- src/pages/Survey/New.tsx
@@ -1,13 +1,0 @@
-import { Box } from '@mui/material';
-import { Flex } from 'components/Box';
-
-const NewSurvey = () => {
-  return (
-    <Box>
-      <Box>asd</Box>
-      <div>asdfa</div>
-    </Box>
-  );
-};
-
-export default NewSurvey;
