+++ src/pages/Survey/New/NewSurveyForm.tsx
@@ -1,9 +1,10 @@
-import { StepProgress } from 'components/Survey';
+import { StepProgress, SurveyCreator } from 'components/Survey';
 
 const NewSurveyForm = () => {
   return (
     <>
       <StepProgress />
+      <SurveyCreator />
       NewSurveyForm
     </>
   );
