+++ src/pages/Survey/New/NewSurveyPayment.tsx
