+++ src/pages/index.tsx
@@ -9,8 +9,8 @@ export { default as SurveyDetail } from 'pages/Survey/SurveyDetail';
 export { default as SurveyResult } from 'pages/Survey/SurveyResult';
 export { default as SurveyGiftResult } from 'pages/Survey/SurveyGiftResult';
 export { default as SurveySubmit } from 'pages/Survey/SurveySubmit';
-export { default as NewSurvey } from 'pages/Survey/NewSurvey';
-export { default as NewSurveyForm } from 'pages/Survey/NewSurveyForm';
-export { default as NewSurveyPayment } from 'pages/Survey/NewSurveyPayment';
-export { default as NewSurveyGiftItem } from 'pages/Survey/NewSurveyGiftItem';
-export { default as NewSurveyGifts } from 'pages/Survey/NewSurveyGifts';
+export { default as NewSurvey } from 'pages/Survey/New/NewSurvey';
+export { default as NewSurveyForm } from 'pages/Survey/New/NewSurveyForm';
+export { default as NewSurveyPayment } from 'pages/Survey/New/NewSurveyPayment';
+export { default as NewSurveyGiftItem } from 'pages/Survey/New/NewSurveyGiftItem';
+export { default as NewSurveyGifts } from 'pages/Survey/New/NewSurveyGifts';
